web/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
celery
redis
amqp
mysql
gunicorn